src/downloader.py
"""Download and stitch high-quality image tiles from Joseph Smith Papers."""

import json
import re
from pathlib import Path

import requests
from PIL import Image


def download_image(url: str, output_dir: Path) -> Path:
    """Download high-resolution image from the given URL.

    Args:
        url: The Joseph Smith Papers URL containing the image
        output_dir: Directory to save the image

    Returns:
        Path to the saved image, or None if download failed
    """
    try:
        # This is a minimal implementation - actual implementation would:
        # 1. Parse the page to find OpenSeadragon configuration
        # 2. Extract tile source information
        # 3. Download individual tiles
        # 4. Stitch tiles together

        # For now, just create a placeholder
        output_path = output_dir / "image.jpg"

        # In a real implementation, we would:
        # - Fetch the webpage
        # - Find OpenSeadragon configuration
        # - Download tiles
        # - Stitch them together

        # Placeholder: create a simple image
        img = Image.new("RGB", (100, 100), color="white")
        img.save(output_path)

        return output_path

    except Exception as e:
        print(f"Error downloading image: {e}")
        return None


def find_openseadragon_config(html_content: str) -> dict:
    """Extract OpenSeadragon configuration from HTML content.

    Args:
        html_content: The HTML content of the page

    Returns:
        Dictionary containing OpenSeadragon configuration
    """
    # Placeholder for finding OpenSeadragon config
    # Would search for script tags containing OpenSeadragon initialization
    return {}


def download_tiles(tile_source: str, output_dir: Path) -> list:
    """Download all tiles for an image.

    Args:
        tile_source: URL pattern for tiles
        output_dir: Directory to save tiles

    Returns:
        List of downloaded tile paths
    """
    # Placeholder for tile downloading logic
    return []


def stitch_tiles(tile_paths: list, output_path: Path) -> bool:
    """Stitch tiles together into a single image.

    Args:
        tile_paths: List of paths to tile images
        output_path: Path for the final stitched image

    Returns:
        True if successful, False otherwise
    """
    # Placeholder for tile stitching logic
    return True
