pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jsp"
version = "1.0.0"
description = "CLI tool for downloading and scraping content from Joseph Smith Papers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["cli", "joseph-smith-papers", "mormon", "scraper", "downloader"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Religion",
    "Topic :: Religion",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "Pillow>=9.0.0",
    "lxml>=4.9.0",
]

[project.urls]
Homepage = "https://github.com/jaredrummler/jsp"
Repository = "https://github.com/jaredrummler/jsp"
Issues = "https://github.com/jaredrummler/jsp/issues"

[project.scripts]
jsp = "src.cli:cli"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  venv
  | build
  | dist
  | .dev
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
fixable = ["ALL"]
unfixable = []
exclude = [
    ".git",
    ".ruff_cache",
    "__pypackages__",
    "build",
    "dist",
    "venv",
    ".dev",
]